move score submission functions to new api
